Changed market data to integer.
Before saving new market data, they get converted into integers.

For some reason, the SQL statements don't work with floats. MariaDB will throw an invalid parameter error, but when executing the SQL command directly via the terminal, the UPDATE statement works perfectly fine. On top of that is the bug inconsistent between the local development environment and the production one. And going even further, the bug does not seem to be consistent on the production server as well. It does not occur always, one time updating the DB server fixed it temporarily. As the floating point problems are only relevant for one item (Tritanium) which is currently not being used by the bot in any calculation, the values will be inserted as integers into the DB.