Reworked config system.
The old config system was simplified to support better modularization. Subconfigs can now easily be created and merged with the main config. This is the first step for the new plugin system.

THIS IS A BREAKING CHANGE, the bot does no longer work after this commit. Only the config.py file was updated. All usages have not been updated yet and will no longer work.