Added plugin lifecycle logic.
All plugins have to be loaded and enabled. Those functions will provide the plugins with entry points to the discord bot to register commands and events (not yet implemented). During startup, the bot searches for an order to load plugins. Every plugin can depend on other plugins, which have to be loaded before the plugin itself can be loaded (cyclic imports are not allowed).
The bot will use this order for loading and enabling. Disabling and unloading will be performed in reverse order.

Every plugin has an on_load and on_enable method (as well as for unloading...), loading is a synchronous process while enabling is asynchronous.