Improved localization and added example.
The bot now gets the localization context from the ApplicationContext during slash commands. This is the language of the discord client from the user.

Also fixed/improved the localization for the example plugin and the help command.

Added plugins.md with more information on how to use the new plugin system.