Added plugin reload command.
Plugins can now be reloaded by the owner with /plugin reload <name>. This will reload the python module from the file system. Imported files will NOT get reloaded, so if the plugin is split over multiple files this does not work.
This command should only be used for testing purposes, it is recommended to perform a full restart for complicated plugins to prevent any unexpected side effects.