Fixed shutdown procedure and added maintenance mode.
When stopping the bot via /bot stop, the plugins got not unloaded properly. Also, plugins can specify menu messages which get edited when shutting down, the bot will remove the view and replace the embed with an info embed. With /bot maintenance it is possible to specify an end time for the maintenance, this time will be shown when the bot gets shut down.