Added embed builder and commands.
The embed builder provides a simple way to build and modify embeds. As opposed to the discord.Embed, all fields may be changed. Once all values are set as desired, the embed can be built using build_embed. There are also new commands (/embed ...) that can modify and post embeds.