Fixed time parsing and improve menu.
The dateutils parser is not able to parse ISO formatted time strings when the dayfirst parameter is set to True, as it tries to parse YYYY-DD-MM. So added a special check for ISO strings.
Added a "Next Month" section to the checklist when in the last week of the current month.